+++ NEWS.md
@@ -36,6 +36,8 @@
   attributes. Part of the implementation of #2064.
 * MKVToolNix GUI: header editor: added support for the "video projection"
   track header attributes. Part of the implementation of #2064.
+* mkvmerge: added support for the "video projection" track header
+  attributes. Part of the implementation of #2064.
 
 ## Bug fixes